+++ src/userflow.ts
@@ -9,6 +9,7 @@ interface WindowWithUserflow extends Window {
   USERFLOWJS_QUEUE?: [string, Deferred | null, any[]][]
 
   USERFLOWJS_ENV_VARS?: Record<string, any>
+  __userflowStatePatched: Boolean
 }
 
 // userflow.js API
@@ -206,6 +207,36 @@ interface Deferred {
 // empty object instead of `window`.
 var w: WindowWithUserflow = typeof window === 'undefined' ? ({} as any) : window
 var userflow = w.userflow
+const history = w.history
+
+// indicates if the history API's pushState and replaceState are patched with custom event emitters
+w.__userflowStatePatched = true
+
+// patch the history API's
+// pushState method to emit userflow:pushstate and
+// replaceState method to emit userflow:replacestate
+const originalPushState = history.pushState
+const originalReplaceState = history.replaceState
+
+function overrideHistoryMethods(method: () => void, eventName: string) {
+  return function () {
+    const event = new CustomEvent(eventName)
+    const args = Array.prototype.slice.call(arguments)
+    const ret = method.apply(history, args as any)
+    w.dispatchEvent(event)
+    return ret
+  }
+}
+
+history.pushState = overrideHistoryMethods(
+  originalPushState as any,
+  'userflow:pushstate'
+)
+history.replaceState = overrideHistoryMethods(
+  originalReplaceState as any,
+  'userflow:replacestate'
+)
+
 if (!userflow) {
   //
   var urlPrefix = 'https://js.userflow.com/'